refactor: goaltype에서 id
