+++ components/TodoItem/TodoItem.tsx
@@ -17,7 +17,7 @@ import PenIcon from '@/public/icon/pen.svg';
 import { CheckedIcon } from '../../app/(workspace)/todoList/_components/CheckedIcon';
 
 import type { TodoType, EditableTodo } from '@/types/todo.type';
-import { Priority } from '@/types/color.type';
+import { Priority } from '@/types/color.types';
 import { useTodoFileDownload } from '@/hooks/todo/useTodoFileDownload';
 
 interface TodoItemProps {
