+++ hooks/goalsDetail/useGetGoalDetail.ts
@@ -4,7 +4,7 @@ import { apiWithClientToken } from '@/apis/clientActions/index';
 import { useInfoStore } from '@/provider/store-provider';
 
 import type { ApiResponse } from '@/types/apiResponse.type';
-import type { GoalType } from '@/types/goal.type';
+import type { GoalType } from '@/types/goal.types';
 import type { TodoType } from '@/types/todo.type';
 
 type GoalDetailResult = GoalType & {
