+++ types/apiResponse.type.ts
@@ -1,4 +1,4 @@
-export type ApiErrorCode = 200 | 400;
+type ApiErrorCode = 200 | 400;
 
 export type ApiResponse<T> = {
   success: boolean;
