+++ types/note.type.ts
@@ -1,6 +1,6 @@
 import type { TodoType } from './todo.type';
 import type { ISODateString } from './ISODateString.type';
-import type { GoalSummary } from './goal.type';
+import type { GoalSummary } from './goal.types';
 export type NoteCommonFields = {
   title: string;
   content: string;
