+++ types/todo.type.ts
@@ -1,6 +1,6 @@
-import { Priority } from './color.type';
-import type { ISODateString } from './goal.type';
-import type { GoalSummary } from './goal.type';
+import { Priority } from './color.types';
+import type { ISODateString } from './goal.types';
+import type { GoalSummary } from './goal.types';
 export type TodoListType = {
   lastSeenId: number;
   remainingCount: number;
