+++ utils/dashboard/goalUtil.ts
@@ -1,9 +1,9 @@
 import { format } from 'date-fns';
 
-import type { GoalId } from '@/types/goal.type';
+import type { GoalType } from '@/types/goal.types';
 import type { TodoType } from '@/types/todo.type';
 
-export const getGoalTodos = (todos: TodoType[], goalId: GoalId['goalId']) => {
+export const getGoalTodos = (todos: TodoType[], goalId: GoalType['id']) => {
   return todos.filter((todo) => todo.goal?.id === goalId);
 };
 
@@ -16,5 +16,5 @@ export const getGoalIdByTodo = (todos: TodoType[]) => {
       }
       return null;
     })
-    .filter(Boolean) as GoalId[];
+    .filter(Boolean) as { goalId: GoalType['id']; completionDate: GoalType['completionDate'] }[];
 };
